+++ libs/adls/src/thds/adls/azcopy/download.py
@@ -7,180 +7,160 @@
 # very end of the download), so for local users who don't have huge bandwidth, it's likely
 # a better user experience to disable this globally.
 import asyncio
-import json
-import os
 import subprocess
 import typing as ty
-import urllib.parse
-from contextlib import contextmanager
+from contextlib import nullcontext
+from dataclasses import dataclass
 from pathlib import Path
 
 from azure.storage.filedatalake import DataLakeFileClient
 
-from thds.core import cache, config, log
+from thds.core import config, log
 
-from .. import _progress, conf, uri
+from .. import conf
+from . import login, progress, system_resources
 
 DONT_USE_AZCOPY = config.item("dont_use", default=False, parse=config.tobool)
-
-_AZCOPY_LOGIN_WORKLOAD_IDENTITY = "azcopy login --login-type workload".split()
-_AZCOPY_LOGIN_LOCAL_STATUS = "azcopy login status".split()
-# device login is an interactive process involving a web browser,
-# which is not acceptable for large scale automation.
-# So instead of logging in, we check to see if you _are_ logged in,
-# and if you are, we try using azcopy in the future.
+MIN_FILE_SIZE = config.item("min_file_size", default=20 * 10**6, parse=int)  # 20 MB
 
 logger = log.getLogger(__name__)
 
 
-class DownloadRequest(ty.NamedTuple):
+@dataclass
+class DownloadRequest:
+    temp_path: Path
+    size_bytes: ty.Optional[int]
+
+
+@dataclass
+class SdkDownloadRequest(DownloadRequest):
     """Use one or the other, but not both, to write the results."""
 
     writer: ty.IO[bytes]
-    temp_path: Path
 
 
-@cache.locking  # only run this once per process.
-def _good_azcopy_login() -> bool:
-    if DONT_USE_AZCOPY():
-        return False
-
-    try:
-        subprocess.run(_AZCOPY_LOGIN_WORKLOAD_IDENTITY, check=True, capture_output=True)
-        logger.info("Will use azcopy for downloads in this process...")
-        return True
-
-    except (subprocess.CalledProcessError, FileNotFoundError):
-        pass
-    try:
-        subprocess.run(_AZCOPY_LOGIN_LOCAL_STATUS, check=True)
-        logger.info("Will use azcopy for downloads in this process...", dl=None)
-        return True
-    except FileNotFoundError:
-        logger.info("azcopy is not installed or not on your PATH, so we cannot speed up downloads")
-    except subprocess.CalledProcessError as cpe:
-        logger.warning(
-            "You are not logged in with azcopy, so we cannot speed up downloads."
-            f" Run `azcopy login` to fix this. Return code was {cpe.returncode}"
-        )
-    return False
+def _is_big_enough_for_azcopy(size_bytes: int) -> bool:
+    return size_bytes >= MIN_FILE_SIZE()
 
 
-def _azcopy_download_command(dl_file_client: DataLakeFileClient, path: Path) -> ty.List[str]:
-    return ["azcopy", "copy", dl_file_client.url, str(path), "--output-type=json"]
-
-
-class AzCopyMessage(ty.TypedDict):
-    TotalBytesEnumerated: str
-    TotalBytesTransferred: str
-
-
-class AzCopyJsonLine(ty.TypedDict):
-    MessageType: str
-    MessageContent: AzCopyMessage
-
-
-def _parse_azcopy_json_output(line: str) -> AzCopyJsonLine:
-    outer_msg = json.loads(line)
-    return AzCopyJsonLine(
-        MessageType=outer_msg["MessageType"],
-        MessageContent=json.loads(outer_msg["MessageContent"]),
+def should_use_azcopy(file_size_bytes: int) -> bool:
+    return (
+        _is_big_enough_for_azcopy(file_size_bytes)
+        and not DONT_USE_AZCOPY()
+        and login.good_azcopy_login()
     )
 
 
-@contextmanager
-def _track_azcopy_progress(http_url: str) -> ty.Iterator[ty.Callable[[str], None]]:
-    """Context manager that tracks progress from AzCopy JSON lines. This works for both async and sync impls."""
-    tracker = _progress.get_global_download_tracker()
-    adls_uri = urllib.parse.unquote(str(uri.parse_uri(http_url)))
-
-    def track(line: str):
-        if not line:
-            return
-
-        try:
-            prog = _parse_azcopy_json_output(line)
-            if prog["MessageType"] == "Progress":
-                tracker(adls_uri, total_written=int(prog["MessageContent"]["TotalBytesTransferred"]))
-        except json.JSONDecodeError:
-            pass
-
-    yield track
-
-
-def _restrict_mem() -> dict:
-    return dict(os.environ, AZCOPY_BUFFER_GB="0.3")
+def _azcopy_download_command(dl_file_client: DataLakeFileClient, path: Path) -> ty.List[str]:
+    # turns out azcopy checks md5 by default - but we we do our own checking, sometimes with faster methods,
+    # and their checking _dramatically_ slows downloads on capable machines, so we disable it.
+    return ["azcopy", "copy", dl_file_client.url, str(path), "--output-type=json", "--check-md5=NoCheck"]
 
 
 def sync_fastpath(
     dl_file_client: DataLakeFileClient,
     download_request: DownloadRequest,
 ) -> None:
-    if _good_azcopy_login():
+    if not isinstance(download_request, SdkDownloadRequest):
+        logger.debug("Downloading %s using azcopy", dl_file_client.url)
         try:
             # Run the copy
             process = subprocess.Popen(
                 _azcopy_download_command(dl_file_client, download_request.temp_path),
                 stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
+                stderr=subprocess.STDOUT,
                 text=True,
-                env=_restrict_mem(),
+                env=system_resources.restrict_usage(),
             )
             assert process.stdout
-            with _track_azcopy_progress(dl_file_client.url) as track:
+            output_lines = list()
+            with progress.azcopy_tracker(
+                "down", dl_file_client.url, download_request.size_bytes or 0
+            ) as track:
                 for line in process.stdout:
                     track(line)
-            return  # success
+                    output_lines.append(line.strip())
+
+            process.wait()
+            if process.returncode != 0:
+                raise subprocess.CalledProcessError(
+                    process.returncode,
+                    f"AzCopy failed with return code {process.returncode}\n\n" + "\n".join(output_lines),
+                )
+            assert (
+                download_request.temp_path.exists()
+            ), f"AzCopy did not create the file at {download_request.temp_path}"
+            return
 
         except (subprocess.SubprocessError, FileNotFoundError):
             logger.warning("Falling back to Python SDK for download")
 
-    dl_file_client.download_file(
-        max_concurrency=conf.DOWNLOAD_FILE_MAX_CONCURRENCY(),
-        connection_timeout=conf.CONNECTION_TIMEOUT(),
-    ).readinto(download_request.writer)
+    logger.debug("Downloading %s using Python SDK", dl_file_client.url)
+    if hasattr(download_request, "writer"):
+        writer_cm = nullcontext(download_request.writer)
+    else:
+        writer_cm = open(download_request.temp_path, "wb")  # type: ignore[assignment]
+    with writer_cm as writer:
+        dl_file_client.download_file(
+            max_concurrency=conf.DOWNLOAD_FILE_MAX_CONCURRENCY(),
+            connection_timeout=conf.CONNECTION_TIMEOUT(),
+        ).readinto(writer)
 
 
 async def async_fastpath(
     dl_file_client: DataLakeFileClient,
     download_request: DownloadRequest,
 ) -> None:
-    # technically it would be 'better' to do this login in an async subproces,
+    # technically it would be 'better' to do this login in an async subprocess,
     # but it involves a lot of boilerplate, _and_ we have no nice way to cache
     # the value, which is going to be computed one per process and never again.
     # So we'll just block the async loop for a couple of seconds one time...
-    if _good_azcopy_login():
+    if not isinstance(download_request, SdkDownloadRequest):
+        logger.debug("Downloading %s using azcopy", dl_file_client.url)
         try:
             # Run the copy
             copy_proc = await asyncio.create_subprocess_exec(
                 *_azcopy_download_command(dl_file_client, download_request.temp_path),
                 stdout=asyncio.subprocess.PIPE,
-                stderr=asyncio.subprocess.PIPE,
-                env=_restrict_mem(),
+                stderr=asyncio.subprocess.STDOUT,
+                env=system_resources.restrict_usage(),
             )
             assert copy_proc.stdout
 
             # Feed lines to the tracker asynchronously
-            with _track_azcopy_progress(dl_file_client.url) as track:
+            output_lines = list()
+            with progress.azcopy_tracker(
+                "down", dl_file_client.url, download_request.size_bytes or 0
+            ) as track:
                 while True:
                     line = await copy_proc.stdout.readline()
                     if not line:  # EOF
                         break
                     track(line.decode().strip())
+                    output_lines.append(line.decode().strip())
 
             # Wait for process completion
             exit_code = await copy_proc.wait()
             if exit_code != 0:
-                raise subprocess.SubprocessError()
+                raise subprocess.CalledProcessError(
+                    exit_code,
+                    f"AzCopy failed with return code {exit_code}\n\n" + "\n".join(output_lines),
+                )
 
-            return  # success
+            return
 
         except (subprocess.SubprocessError, FileNotFoundError):
             logger.warning("Falling back to Python SDK for download")
 
-    reader = await dl_file_client.download_file(  # type: ignore[misc]
-        # TODO - check above type ignore
-        max_concurrency=conf.DOWNLOAD_FILE_MAX_CONCURRENCY(),
-        connection_timeout=conf.CONNECTION_TIMEOUT(),
-    )
-    await reader.readinto(download_request.writer)
+    logger.debug("Downloading %s using Async Python SDK", dl_file_client.url)
+    if hasattr(download_request, "writer"):
+        writer_cm = nullcontext(download_request.writer)
+    else:
+        writer_cm = open(download_request.temp_path, "wb")  # type: ignore[assignment]
+    with writer_cm as writer:
+        reader = await dl_file_client.download_file(  # type: ignore[misc]
+            # TODO - check above type ignore
+            max_concurrency=conf.DOWNLOAD_FILE_MAX_CONCURRENCY(),
+            connection_timeout=conf.CONNECTION_TIMEOUT(),
+        )
+        await reader.readinto(writer)