+++ libs/adls/src/thds/adls/tools/download.py
@@ -1,7 +1,7 @@
 import argparse
 from pathlib import Path
 
-from thds.adls.cached_up_down import download_directory, download_to_cache
+from thds.adls import cached
 from thds.adls.file_properties import get_file_properties, is_directory
 from thds.adls.impl import ADLSFileSystem
 from thds.adls.uri import resolve_uri
@@ -16,9 +16,10 @@ def main():
         help="A fully qualified path to an ADLS location. Accepts adls://, https:// and abfss:// URIs.",
     )
     parser.add_argument(
-        "--copy-to",
-        "-c",
+        "copy_to",
+        nargs="?",
         type=Path,
+        default=None,
         help="This will create a link to the cached download at the specified location",
     )
     parser.add_argument(
@@ -39,9 +40,9 @@ def main():
             cache_path = fs.fetch_file(args.adls_fqn.path)
     else:
         if is_dir:
-            cache_path = download_directory(args.adls_fqn)
+            cache_path = cached.download_directory(args.adls_fqn)
         else:
-            cache_path = download_to_cache(args.adls_fqn)
+            cache_path = cached.download_to_cache(args.adls_fqn)
 
     if args.copy_to:
         link(cache_path, args.copy_to)