+++ libs/adls/tests/test_adls/test_download.py
@@ -9,15 +9,12 @@ import pytest
 from azure.identity.aio import DefaultAzureCredential
 from azure.storage.filedatalake import FileProperties, FileSystemClient, aio
 
-from thds.adls import ADLSFileSystem, AdlsFqn, AdlsRoot, azcopy
-from thds.adls.cached_up_down import download_to_cache
+from thds import adls
+from thds.adls import ADLSFileSystem, AdlsFqn, AdlsRoot, azcopy, cached, errors, hashes, md5
 from thds.adls.download import (
-    MD5MismatchError,
     _download_or_use_verified_cached_coroutine,
     _IoRequest,
-    _verify_md5s_before_and_after_download,
     async_download_or_use_verified,
-    b64,
     download_or_use_verified,
 )
 from thds.adls.download_lock import _clean_download_locks
@@ -43,13 +40,13 @@ def test_unit_download_coroutine_no_cache_no_remote_md5b64(test_dest: Path):
     assert request == _IoRequest.FILE_PROPERTIES
     # this time, we're asking because we don't have an md5
 
-    request = co.send(FileProperties())
+    request = co.send(FileProperties(name="exist.lol"))
     assert request == _IoRequest.FILE_PROPERTIES
     # this time, we're asking because we're about to do a download and
     # we need to know whether to skip the download.
 
-    wfp = co.send(FileProperties())
-    assert isinstance(wfp, azcopy.download.DownloadRequest)
+    wfp = co.send(FileProperties(name="exist.lol"))
+    assert isinstance(wfp, azcopy.download.SdkDownloadRequest)
     wfp.writer.write(b"ello")
 
     with pytest.raises(StopIteration) as si:
@@ -67,14 +64,14 @@ def test_integration_download_to_local_and_reuse_from_there(
     fqn = AdlsFqn(
         ty.cast(str, global_test_fs_client.account_name), global_test_fs_client.file_system_name, real
     )
-    md5b64 = "U3vtigRGuroWtJFEQ5dKoQ=="
+    the_hash = md5.to_hash(md5b64="U3vtigRGuroWtJFEQ5dKoQ==")
     lcl = test_dest / "DONT_DELETE_THESE_FILES.txt"
-    hit = download_or_use_verified(global_test_fs_client, real, lcl, md5b64)
+    hit = download_or_use_verified(global_test_fs_client, real, lcl, expected_hash=the_hash)
     assert not hit
     assert lcl.exists()
     assert not _TEST_CACHE.path(fqn).exists()
 
-    hit = download_or_use_verified(global_test_fs_client, real, lcl, md5b64)
+    hit = download_or_use_verified(global_test_fs_client, real, lcl, expected_hash=the_hash)
     assert hit
     assert lcl.exists()
     assert not _TEST_CACHE.path(fqn).exists()
@@ -87,20 +84,26 @@ def test_integration_download_to_cache_with_no_expected_md5_and_reuse_from_there
     fqn = AdlsFqn(
         ty.cast(str, global_test_fs_client.account_name), global_test_fs_client.file_system_name, remote
     )
-    md5b64 = ""
     lcl = test_dest / "DONT_DELETE_THESE_FILES----use-cache.txt"
-    hit = download_or_use_verified(global_test_fs_client, remote, lcl, md5b64, cache=_TEST_CACHE)
+    hit = download_or_use_verified(
+        global_test_fs_client, remote, lcl, expected_hash=None, cache=_TEST_CACHE
+    )
     assert not hit
     assert lcl.exists()
     assert _TEST_CACHE.path(fqn).exists()
 
-    hit = download_or_use_verified(global_test_fs_client, remote, lcl, md5b64, cache=_TEST_CACHE)
+    hit = download_or_use_verified(
+        global_test_fs_client, remote, lcl, expected_hash=None, cache=_TEST_CACHE
+    )
     assert hit
     assert lcl.exists()
+    assert hit.exists()
     assert _TEST_CACHE.path(fqn).exists()
 
     newlcl = test_dest / "DONT_DELETE---put-in-different-place-but-use-cache.txt"
-    hit = download_or_use_verified(global_test_fs_client, remote, newlcl, md5b64, cache=_TEST_CACHE)
+    hit = download_or_use_verified(
+        global_test_fs_client, remote, newlcl, expected_hash=None, cache=_TEST_CACHE
+    )
     assert newlcl.exists()
     assert lcl.exists()  # still...
     assert _TEST_CACHE.path(fqn).exists()
@@ -116,9 +119,11 @@ def test_integration_handles_emoji_and_long_key(
     fqn = AdlsFqn(
         ty.cast(str, global_test_fs_client.account_name), global_test_fs_client.file_system_name, remote
     )
-    md5b64 = "gEL83AfKoP2e3O1Y4RsBqQ=="
+    md5h = md5.to_hash(md5b64="gEL83AfKoP2e3O1Y4RsBqQ==")
     lcl = test_dest / "benchmark_hashing.py"
-    hit = download_or_use_verified(global_test_fs_client, remote, lcl, md5b64, cache=_TEST_CACHE)
+    hit = download_or_use_verified(
+        global_test_fs_client, remote, lcl, expected_hash=md5h, cache=_TEST_CACHE
+    )
     assert not hit
     assert lcl.exists()
     assert _TEST_CACHE.path(fqn).exists()
@@ -127,21 +132,22 @@ def test_integration_handles_emoji_and_long_key(
 
 def test_integration_md5_verification(global_test_fs_client: FileSystemClient, test_dest: Path):
     real = "test/read-only/DONT_DELETE_THESE_FILES.txt"
-    md5b64 = "incorrect-MrMjF87w3GvA=="
+    md5h = md5.to_hash(md5b64="incorrect-MrMjF87w3GvA1==")
     lcl = test_dest / "DONT_DELETE_THESE_FILES.txt"
-    with pytest.raises(MD5MismatchError):
-        download_or_use_verified(global_test_fs_client, real, lcl, md5b64)
+    with pytest.raises(errors.HashMismatchError):
+        download_or_use_verified(global_test_fs_client, real, lcl, expected_hash=md5h)
 
 
 def test_unit_md5_verification(test_dest: Path):
     made_it = False
-    with pytest.raises(MD5MismatchError):
+    with pytest.raises(errors.HashMismatchError):
         local_dest = test_dest / "a-file.txt"
         with open(local_dest, "w") as f:
             f.write("hi")
-        with _verify_md5s_before_and_after_download(
-            "WPMVPiXYwhMrMjF87w3GvA==",
-            "WPMVPiXYwhMrMjF87w3GvA==",
+        md5h = md5.to_hash(md5b64="WPMVPiXYwhMrMjF87w3GvA==")
+        with hashes.verify_hashes_before_and_after_download(
+            md5h,
+            md5h,
             AdlsFqn("foo", "bar", "baz"),
             local_dest,
         ):
@@ -162,17 +168,19 @@ async def test_integration_async(test_remote_root: AdlsRoot, test_dest: Path):
     lcl = test_dest / "DONT_DELETE_THESE_FILES----use-async.txt"
 
     async_client = _async_adls_fs_client(*test_remote_root)
-    hit = await async_download_or_use_verified(async_client, remote, lcl, "U3vtigRGuroWtJFEQ5dKoQ==")
+    md5h = md5.to_hash(md5b64="U3vtigRGuroWtJFEQ5dKoQ==")
+    hit = await async_download_or_use_verified(async_client, remote, lcl, expected_hash=md5h)
     assert not hit
     assert lcl.exists()
 
-    hit = await async_download_or_use_verified(async_client, remote, lcl, "U3vtigRGuroWtJFEQ5dKoQ==")
+    hit = await async_download_or_use_verified(async_client, remote, lcl, expected_hash=md5h)
     assert hit
     assert lcl.exists()
+    assert hit.exists()
 
 
 @pytest.mark.asyncio
-async def test_file_missing_md5_gets_one_assigned_after_download(
+async def test_file_missing_hash_gets_one_assigned_after_download(
     tmp_remote_root: AdlsRoot, test_dest: Path
 ):
     fs_client = _async_adls_fs_client(*tmp_remote_root)
@@ -186,8 +194,9 @@ async def test_file_missing_md5_gets_one_assigned_after_download(
     assert not cache_hit
 
     fp = await file_client.get_file_properties()
-    assert fp.content_settings.content_md5
-    assert b64(fp.content_settings.content_md5) == "8Wz15VCq6d73Z0+KUDNqVg=="
+    assert not fp.content_settings.content_md5
+    assert fp.metadata
+    assert fp.metadata["hash_xxh3_128_b64"] == "hOfn6RuS0zmzAhyh1tid6w=="
 
     # should not error since the md5 should be correct
     cache_hit = await async_download_or_use_verified(fs_client, key, test_dest / "missing-md5.txt")
@@ -220,7 +229,8 @@ def random_test_file_fqn(
     with open(random_file, "w") as f:
         f.write(uuid4().hex)
 
-    fs.put_file(random_file, random_test_file_path)  # non-cached upload
+    adls.upload(fqn, random_file, content_type="text/plain")
+    # this upload is not cached but it _does_ compute the standard preferred hash.
 
     yield fqn
 
@@ -236,14 +246,14 @@ def test_parallel_downloads_only_perform_a_single_download(
         # so that we can see them in the test is to use threads
         # so that everything shares the same logging config.
         with concurrent.futures.ThreadPoolExecutor() as executor:
-            list(executor.map(download_to_cache, [random_test_file_fqn] * 10))
+            list(executor.map(cached.download_to_cache, [random_test_file_fqn] * 10))
 
     download_count = 0
     reuse_count = 0
     for record in caplog.records:
         if "Downloading" in record.getMessage():
             download_count += 1
-        elif "Local path matches MD5" in record.getMessage():
+        elif "Local path matches" in record.getMessage():
             reuse_count += 1
 
     global_cache().path(random_test_file_fqn).unlink()
@@ -270,8 +280,8 @@ def test_dont_use_azcopy_if_present(global_test_fs_client: FileSystemClient, tes
         # then all the integration tests above are testing the fallback code anyway.
         # Whatever. we just want to make sure all the code paths are tested wherever they can be.
         real = "test/read-only/DONT_DELETE_THESE_FILES.txt"
-        md5b64 = "U3vtigRGuroWtJFEQ5dKoQ=="
+        md5h = md5.to_hash(md5b64="U3vtigRGuroWtJFEQ5dKoQ==")
         lcl = test_dest / "dont_use_azcopy_DONT_DELETE_THESE_FILES.txt"
-        hit = download_or_use_verified(global_test_fs_client, real, lcl, md5b64)
+        hit = download_or_use_verified(global_test_fs_client, real, lcl, expected_hash=md5h)
         assert not hit
         assert lcl.exists()