+++ libs/core/src/thds/core/source/serde.py
@@ -1,7 +1,6 @@
 # this should later get promoted somewhere, probably
 import json
 import typing as ty
-from functools import partial
 from pathlib import Path
 
 from thds.core import files, hashing, log, types
@@ -11,24 +10,21 @@ from .src import Source
 
 _SHA256_B64 = "sha256b64"
 _MD5_B64 = "md5b64"
+MD5 = "md5"
 
 logger = log.getLogger(__name__)
 
 
-def _from_sha256b64(d: dict) -> ty.Optional[hashing.Hash]:
-    if "sha256b64" in d:
-        return hashing.Hash(algo="sha256", bytes=hashing.db64(d[_SHA256_B64]))
-    return None
-
-
-def _from_md5b64(d: dict) -> ty.Optional[hashing.Hash]:
-    if "md5b64" in d:
-        return hashing.Hash(algo="md5", bytes=hashing.db64(d[_MD5_B64]))
+def _from_b64(d: dict) -> ty.Optional[hashing.Hash]:
+    for key in d.keys():
+        if key.endswith("b64"):
+            algo = key[:-3]
+            return hashing.Hash(algo=algo, bytes=hashing.db64(d[key]))
     return None
 
 
 HashParser = ty.Callable[[dict], ty.Optional[hashing.Hash]]
-_BASE_PARSERS = (_from_sha256b64, _from_md5b64)
+_BASE_PARSERS = (_from_b64,)
 
 
 def base_parsers() -> ty.Tuple[HashParser, ...]:
@@ -43,19 +39,12 @@ def from_json(json_source: str, hash_parsers: ty.Collection[HashParser] = base_p
     )
 
 
-def _generic_hash_serializer(
-    algo: str, stringify_hash: ty.Callable[[bytes], str], keyname: str, hash: hashing.Hash
-) -> ty.Optional[dict]:
-    if hash.algo == algo:
-        return {keyname: stringify_hash(hash.bytes)}
-    return None
+def _very_generic_b64_hash_serializer(hash: hashing.Hash) -> dict:
+    return {hash.algo + "b64": hashing.b64(hash.bytes)}
 
-
-_to_sha256b64 = partial(_generic_hash_serializer, "sha256", hashing.b64, _SHA256_B64)
-_to_md5b64 = partial(_generic_hash_serializer, "md5", hashing.b64, _MD5_B64)
 
 HashSerializer = ty.Callable[[hashing.Hash], ty.Optional[dict]]
-_BASE_HASH_SERIALIZERS: ty.Tuple[HashSerializer, ...] = (_to_md5b64, _to_sha256b64)  # type: ignore
+_BASE_HASH_SERIALIZERS: ty.Tuple[HashSerializer, ...] = (_very_generic_b64_hash_serializer,)
 
 
 def base_hash_serializers() -> ty.Tuple[HashSerializer, ...]: