+++ libs/mops/src/thds/mops/pure/core/lock/maintain.py
@@ -15,12 +15,14 @@ from datetime import datetime, timedelta
 from functools import partial
 from threading import Thread
 
-from thds.core import log
+from thds.core import config, log
 
 from ._funcs import make_lock_uri
 from .read import get_writer_id, make_read_lockfile
 from .types import LockAcquired
-from .write import LockfileWriter, make_lock_contents
+from .write import LockEmitter, LockfileWriter
+
+MAINTAIN_LOCKS = config.item("thds.mops.pure.local.maintain_locks", default=True, parse=config.tobool)
 
 logger = log.getLogger(__name__)
 
@@ -103,7 +105,7 @@ def remote_lock_maintain(lock_dir_uri: str, expected_writer_id: str = "") -> Loc
     lockfile_writer = LockfileWriter(
         current_writer_id,
         lock_dir_uri,
-        make_lock_contents(get_writer_id(lock_contents), timedelta(seconds=expire_s)),
+        LockEmitter(get_writer_id(lock_contents), timedelta(seconds=expire_s)),
         expire_s,
         writer_name="remote",
     )
@@ -148,3 +150,20 @@ def launch_daemon_lock_maintainer(lock_acq: LockAcquired) -> ty.Callable[[], Non
         lock_acq.release()
 
     return stop_maintaining
+
+
+def maintain_to_release(
+    acquired_lock: LockAcquired,
+) -> ty.Callable[[], None]:
+    """Depending on configuration, potentially start maintaining the lock.
+
+    Return a callable that will release the lock when called.
+    """
+    if MAINTAIN_LOCKS():
+        return launch_daemon_lock_maintainer(acquired_lock)
+
+    return acquired_lock.release
+
+
+def no_maintain() -> None:
+    MAINTAIN_LOCKS.set_global(False)